Crate pcre2

This crate provides a safe high level Rust binding to PCRE2.

The API of this crate attempts to correspond closely to the API of Rust’s regex crate. The API provided by this crate neither matches the full API of Rust’s regex crate nor does it expose the full functionality of PCRE2. Contributions are welcome to improve this.

Modules

bytes

PCRE2 regular expressions for matching on arbitrary bytes.

Structs

Error

A PCRE2 error.

Enums

ErrorKind

The kind of an error indicates the type of operation that was attempted that resulted in an error.

Functions

is_jit_available

Returns true if and only if PCRE2 believes that JIT is available.

version

Returns the version of PCRE2 being used.